backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langgraph==0.0.15
chromadb==0.4.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-genai
redis>=4.0.0